main/api.py
from rest_framework import viewsets

from main.models import Item
from main.serializers import ItemSerializer


class ItemViewSet(viewsets.ModelViewSet):
    queryset = Item.objects.all()
    serializer_class = ItemSerializer
