requirements.txt
amqp==2.6.1
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.7
certifi==2020.6.20
coverage==5.3
Django==3.1.1
django-appconf==1.0.4
django-celery-email==3.0.0
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.3.0
django-phonenumber-field==5.0.0
djangorestframework==3.11.1
kombu==4.6.11
Markdown==3.2.2
mysqlclient==2.0.1
phonenumbers==8.12.9
Pillow==8.0.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
sentry-sdk==0.17.3
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
